requirements.txt
# Lead Automation Requirements
# ============================

# Google APIs
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
gspread==6.0.0

# HTTP requests
requests==2.31.0
httpx==0.26.0

# Anthropic Claude API
anthropic==0.40.0

# Web scraping (for company research)
beautifulsoup4==4.12.3
lxml==5.1.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Scheduling
schedule==1.2.1
apscheduler==3.10.4

# Data handling
pandas==2.1.4

# Utilities
tenacity==8.2.3  # Retry logic
python-dateutil==2.8.2

# Logging
structlog==24.1.0
